requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Document processing
pypdf==3.17.1
python-docx==1.1.0
langchain==0.1.0
langchain-text-splitters==0.0.1

# Vector database
chromadb==0.4.18

# Embeddings and LLM
openai==1.3.5
tiktoken==0.5.2
sentence-transformers==2.2.2

# Web interface
streamlit==1.28.2
streamlit-chat==0.1.1

# Utilities
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
pydantic==2.5.0